fix: adjust import and config file
