+++ src/config/payrollAgreement/payrollAgreementTab/forms/cyclespaymentLabels/index.ts
@@ -1,0 +1,13 @@
+const cyclespaymentLabels = {
+  sendButton: "Enviar",
+  nextButton: "Siguiente",
+  previousButton: "Anterior",
+  cancelButton: "Cancelar",
+  addPaymentCycle: "Agregar ciclo de pago",
+  titlePaymentCycle: "Agregar ciclo de pago",
+  emptyDataMessage:
+    "Aún no hay ningún ciclo de pago registrado, presiona “+ Agregar ciclo de pago” para empezar.",
+  actionText: "Agregar",
+};
+
+export { cyclespaymentLabels };
