+++ src/design/forms/extraordinaryPaymentCycles/index.tsx
@@ -3,24 +3,8 @@ import { FormikProps } from "formik";
 
 import { IExtraordinaryCyclesEntry } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IExtraordinaryCyclesEntry";
 import { useExtraordinaryCyclesForm } from "@hooks/payrollAgreement/useExtraordinaryCyclesForm";
-import { IOrdinaryCyclesEntry } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IOrdinaryCyclesEntry";
+import { IExtraordinaryPaymentCyclesForm } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IExtraordinaryPaymentCyclesForm";
 import { ExtraordinaryPaymentCyclesFormUI } from "./interface";
-
-interface IExtraordinaryPaymentCyclesForm {
-  extraordinaryPayment: IExtraordinaryCyclesEntry[];
-  typeRegularPayroll: boolean;
-  regularPaymentCycles: IOrdinaryCyclesEntry[];
-  setExtraordinaryPayment: React.Dispatch<
-    React.SetStateAction<IExtraordinaryCyclesEntry[]>
-  >;
-  onButtonClick: () => void;
-  onPreviousStep: () => void;
-  loading?: boolean;
-  onFormValid?: React.Dispatch<React.SetStateAction<boolean>>;
-  onSubmit?: (values: IExtraordinaryCyclesEntry) => void;
-  editDataOption?: boolean;
-  initialData?: IExtraordinaryCyclesEntry[];
-}
 
 const ExtraordinaryPaymentCyclesForm = forwardRef<
   FormikProps<IExtraordinaryCyclesEntry>,
