+++ src/design/forms/extraordinaryPaymentCycles/interface.tsx
@@ -1,45 +1,21 @@
 import { MdOutlineAdd } from "react-icons/md";
-import { FormikProps } from "formik";
 import { Button, Stack } from "@inubekit/inubekit";
-
 import { tokens } from "@design/tokens";
 import { Table } from "@design/data/table";
 import { ComponentAppearance } from "@enum/appearances";
-import { IEntry } from "@design/data/table/types";
 import {
   breakPoints,
   titles,
   actionsConfig,
 } from "@config/payrollAgreement/payrollAgreementTab/assisted/extraordinaryCyclesTable";
-import { IExtraordinaryCyclesEntry } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IExtraordinaryCyclesEntry";
-import { IServerDomain } from "@ptypes/IServerDomain";
 import { AddCycleModal } from "@design/modals/addCycleModal";
+import { cyclespaymentLabels } from "@config/payrollAgreement/payrollAgreementTab/forms/cyclespaymentLabels";
+import { IExtraordinaryPaymentCyclesFormUI } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IExtraordinaryPaymentCyclesFormUI";
 import {
   StyledContainer,
   StyledContainerFields,
   StyledFormContent,
 } from "./styles";
-
-interface IExtraordinaryPaymentCyclesFormUI {
-  formik: FormikProps<IExtraordinaryCyclesEntry>;
-  entries: IEntry[];
-  editDataOption: boolean;
-  valuesEqual: boolean;
-  loading: boolean;
-  showModal: boolean;
-  isMobile: boolean;
-  typePaymentOptions: IServerDomain[];
-  numberDaysUntilCutOptions: IServerDomain[];
-  monthOptions: IServerDomain[];
-  dayOptions: IServerDomain[];
-  onAddCycle: () => void;
-  onToggleModal: () => void;
-  onButtonClick: () => void;
-  onChange: (name: string, value: string) => void;
-  onPreviousStep: () => void;
-  setEntryDeleted: (value: string | number) => void;
-  isDisabledButton?: boolean;
-}
 
 const ExtraordinaryPaymentCyclesFormUI = (
   props: IExtraordinaryPaymentCyclesFormUI,
@@ -82,7 +58,7 @@ const ExtraordinaryPaymentCyclesFormUI = (
                 onClick={onToggleModal}
                 appearance={ComponentAppearance.PRIMARY}
               >
-                Agregar ciclo de pago
+                {cyclespaymentLabels.titlePaymentCycle}
               </Button>
 
               <Table
@@ -94,7 +70,7 @@ const ExtraordinaryPaymentCyclesFormUI = (
                 isLoading={loading}
                 columnWidths={[50, 12, 10, 14]}
                 withActionsTitles
-                emptyDataMessage="Aún no hay ningún ciclo de pago registrado, presiona “+ Agregar ciclo de pago” para empezar."
+                emptyDataMessage={cyclespaymentLabels.emptyDataMessage}
               />
             </Stack>
           </StyledContainerFields>
@@ -107,7 +83,9 @@ const ExtraordinaryPaymentCyclesFormUI = (
           appearance={ComponentAppearance.GRAY}
           variant="outlined"
         >
-          {editDataOption ? "Cancelar" : "Anterior"}
+          {editDataOption
+            ? cyclespaymentLabels.cancelButton
+            : cyclespaymentLabels.previousButton}
         </Button>
 
         <Button
@@ -117,17 +95,19 @@ const ExtraordinaryPaymentCyclesFormUI = (
           loading={loading}
           appearance={ComponentAppearance.PRIMARY}
         >
-          {editDataOption ? "Enviar" : "Siguiente"}
+          {editDataOption
+            ? cyclespaymentLabels.sendButton
+            : cyclespaymentLabels.nextButton}
         </Button>
       </Stack>
       {showModal && (
         <AddCycleModal
-          actionText="Agregar"
+          actionText={cyclespaymentLabels.actionText}
           comparisonData={valuesEqual}
           formik={formik}
           isLoading={loading}
           portalId="portal"
-          title="Agregar ciclo de pago"
+          title={cyclespaymentLabels.addPaymentCycle}
           isExtraordinary
           typePaymentOptions={typePaymentOptions}
           monthOptions={monthOptions}
