+++ src/design/forms/regularPaymentCycles/index.tsx
@@ -4,21 +4,8 @@ import { FormikProps } from "formik";
 import { IOrdinaryCyclesEntry } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IOrdinaryCyclesEntry";
 import { useOrdinaryCyclesForm } from "@hooks/payrollAgreement/useOrdinaryCyclesForm";
 import { infoPeriodicityModal } from "@config/payrollAgreement/payrollAgreementTab/generic/infoPeriodicityModal";
+import { IRegularPaymentCyclesForm } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IRegularPaymentCyclesForm";
 import { RegularPaymentCyclesFormUI } from "./interface";
-
-interface IRegularPaymentCyclesForm {
-  regularPaymentCycles: IOrdinaryCyclesEntry[];
-  onButtonClick: () => void;
-  onPreviousStep: () => void;
-  setRegularPaymentCycles: React.Dispatch<
-    React.SetStateAction<IOrdinaryCyclesEntry[]>
-  >;
-  loading?: boolean;
-  onFormValid?: React.Dispatch<React.SetStateAction<boolean>>;
-  onSubmit?: (values: IOrdinaryCyclesEntry) => void;
-  editDataOption?: boolean;
-  initialData?: IOrdinaryCyclesEntry[];
-}
 
 const RegularPaymentCyclesForm = forwardRef<
   FormikProps<IOrdinaryCyclesEntry>,
