+++ src/types/payrollAgreement/payrollAgreementTab/forms/IExtraordinaryPaymentCyclesForm/index.ts
@@ -1,0 +1,20 @@
+import { IExtraordinaryCyclesEntry } from "../IExtraordinaryCyclesEntry";
+import { IOrdinaryCyclesEntry } from "../IOrdinaryCyclesEntry";
+
+interface IExtraordinaryPaymentCyclesForm {
+  extraordinaryPayment: IExtraordinaryCyclesEntry[];
+  typeRegularPayroll: boolean;
+  regularPaymentCycles: IOrdinaryCyclesEntry[];
+  setExtraordinaryPayment: React.Dispatch<
+    React.SetStateAction<IExtraordinaryCyclesEntry[]>
+  >;
+  onButtonClick: () => void;
+  onPreviousStep: () => void;
+  loading?: boolean;
+  onFormValid?: React.Dispatch<React.SetStateAction<boolean>>;
+  onSubmit?: (values: IExtraordinaryCyclesEntry) => void;
+  editDataOption?: boolean;
+  initialData?: IExtraordinaryCyclesEntry[];
+}
+
+export type { IExtraordinaryPaymentCyclesForm };
