+++ src/types/payrollAgreement/payrollAgreementTab/forms/IRegularPaymentCyclesForm/index.ts
@@ -1,0 +1,17 @@
+import { IOrdinaryCyclesEntry } from "../IOrdinaryCyclesEntry";
+
+interface IRegularPaymentCyclesForm {
+  regularPaymentCycles: IOrdinaryCyclesEntry[];
+  onButtonClick: () => void;
+  onPreviousStep: () => void;
+  setRegularPaymentCycles: React.Dispatch<
+    React.SetStateAction<IOrdinaryCyclesEntry[]>
+  >;
+  loading?: boolean;
+  onFormValid?: React.Dispatch<React.SetStateAction<boolean>>;
+  onSubmit?: (values: IOrdinaryCyclesEntry) => void;
+  editDataOption?: boolean;
+  initialData?: IOrdinaryCyclesEntry[];
+}
+
+export type { IRegularPaymentCyclesForm };
